+++ tests/testthat/test-deprecated.R
@@ -13,3 +13,8 @@ test_that("random_encourage() is deprecated", {
 test_that("grade_learnr() is deprecated", {
   lifecycle::expect_deprecated(grade_learnr())
 })
+
+test_that("fail_code_feedback= is deprecated", {
+  lifecycle::expect_deprecated(gradethis_setup(fail_code_feedback = FALSE))
+  lifecycle::expect_deprecated(grade_this(fail(), fail_code_feedback = FALSE))
+})